refactor(evals): fold code-eval save helpers into utils/evals; slim the V2 view branches
Consolidates parse_evaluate_params, build_function_params_schema,
build_code_eval_config, and refresh_code_eval_schema into
model_hub/utils/evals.py alongside prepare_user_eval_config. Drops the
separate code_eval_signature module. EvalTemplateCreateV2View and
EvalTemplateUpdateView code branches now call one helper each; the
inline signature-derivation, config-dict build, and duplicated import
blocks all live behind that helper.
